+++ src/components/molecules/Input/index.tsx
@@ -1,10 +1,19 @@
 import React, {forwardRef, useState, useRef} from 'react';
-import {StyleSheet, TextInput, View, TouchableWithoutFeedback, Keyboard} from 'react-native';
+import {
+	StyleSheet,
+	TextInput,
+	View,
+	TouchableWithoutFeedback,
+	Keyboard,
+	Platform,
+} from 'react-native';
 import BaseInput, {BaseInputProps} from 'atoms/BaseInput';
 import {palette} from 'theme/palette';
 import {moderateScale, scaledForDevice} from 'scale';
 import handleChangeText from './utils/handleChangeText';
 import Typography from 'atoms/Typography';
+
+const isWeb = Platform.OS === 'web';
 
 enum InputType {
 	currency = 'numeric',
@@ -25,6 +34,7 @@ interface BaseInputPropsExtended extends BaseInputProps {
 	variant?: InputVariant;
 	onChangeText?: (text: string) => void;
 	totalValue?: number;
+	placeholder?: string;
 }
 
 type AmountTotalProps = BaseInputPropsExtended & {
@@ -46,8 +56,9 @@ type DefaultProps = BaseInputPropsExtended & {
 export type InputProps = AmountTotalProps | OtherVariantProps | DefaultProps;
 
 const Input = forwardRef<TextInput, InputProps>(
-	({style, type, variant = 'default', totalValue, onChangeText, ...props}, ref) => {
+	({style, type, variant = 'default', totalValue, placeholder, onChangeText, ...props}, ref) => {
 		const [value, setValue] = useState('');
+		const isPlaceholderBeingShown = !value;
 		const isAmountTotalVariant = variant === 'amountTotal';
 
 		const internalRef = useRef<TextInput>(null);
@@ -68,13 +79,26 @@ const Input = forwardRef<TextInput, InputProps>(
 				justifyContent: 'center',
 				alignItems: 'center',
 				flexDirection: 'row',
+				position: 'relative',
+				width: '100%',
 			},
 			input: {
 				color: palette.black.main,
 				fontSize: scaledForDevice(42, moderateScale),
+				height: '100%',
+				textAlign: 'center',
+				textAlignVertical: 'center',
+				includeFontPadding: false,
+				paddingVertical: 0,
+				...(isPlaceholderBeingShown && {marginLeft: scaledForDevice(-12, moderateScale)}),
+				...(isWeb && {flex: 1, maxWidth: isPlaceholderBeingShown ? '1%' : undefined}),
 			},
 			totalValue: {
 				color: palette.primary.main,
+			},
+			placeholder: {
+				color: '#A8AAAC',
+				marginLeft: scaledForDevice(-12, moderateScale),
 			},
 		});
 
@@ -104,19 +128,33 @@ const Input = forwardRef<TextInput, InputProps>(
 			return InputType.default;
 		})();
 
+		const renderPlaceholder = () => {
+			if (value.length > 0) {
+				return null;
+			}
+
+			return (
+				<Typography type="display" style={styles.placeholder}>
+					{placeholder}
+				</Typography>
+			);
+		};
+
 		return (
 			<TouchableWithoutFeedback onPress={handlePress}>
-				<View style={styles.container}>
+				<View style={[styles.container, style]}>
 					<BaseInput
-						style={[styles.input, style]}
+						testID="input"
+						style={styles.input}
 						ref={inputRef}
 						value={value}
 						keyboardType={resolvedKeyboardType}
 						onChangeText={changeTextCb}
 						{...props}
 					/>
+					{renderPlaceholder()}
 					{isAmountTotalVariant && (
-						<Typography style={styles.totalValue} type="display" size="medium">
+						<Typography style={styles.totalValue} type="display">
 							{`/${totalValue?.toString()}`}
 						</Typography>
 					)}